move project creation to /project
